fix stringobject, add to unicodeobject, setobject (while discovering stubs.py)

+++ pypy/module/cpyext/object.py
@@ -387,7 +387,7 @@ def PyObject_Hash(space, w_obj):
     This is the equivalent of the Python expression hash(o)."""
     return space.int_w(space.hash(w_obj))
 
-@cpython_api([rffi.LONG], rffi.DOUBLE, error=-1)
+@cpython_api([rffi.DOUBLE], rffi.LONG, error=-1)
 def _Py_HashDouble(space, w_obj):
     raise OperationError(space.w_NotImplementedError, 
                 space.wrap("_Py_HashDouble not implemented yet"))
